api/requirements.txt
connexion~=2.7.0
flask~=1.1.2
flask-sqlalchemy
jsonschema~=3.2.0
sqlalchemy
swagger-ui-bundle
PyJWT=2.4.0
markupsafe==2.0.1
transformers
